feat: docker-compose working for sidecar + core-node + postgres

+++ stacks-blockchain/dist-tool/index.js
@@ -4,17 +4,18 @@ const path = require('path');
 const util = require('util');
 const readline = require('readline');
 const { PassThrough, pipeline } = require('stream');
-const { execSync, spawn } = require('child_process');
+const { execSync, execFileSync, spawn } = require('child_process');
 const { Storage } = require('@google-cloud/storage');
 const deepmerge = require('deepmerge');
+const sanitize = require('sanitize-filename');
 
 const pipelineAsync = util.promisify(pipeline);
 
 const BUCKET_NAME = 'blockstack-stacks-blockchain_artifacts';
 const BUCKET_URL = `https://${BUCKET_NAME}.storage.googleapis.com/`;
-const STACKS_BLOCKCHAIN_REPO = 'https://github.com/blockstack/stacks-blockchain.git';
 
 const envVars = {
+  STACKS_BLOCKCHAIN_REPO: 'https://github.com/blockstack/stacks-blockchain.git',
   STACKS_BLOCKCHAIN_BRANCH: 'master',
   STACKS_BLOCKCHAIN_BIN: 'blockstack-core',
   STACKS_BLOCKCHAIN_DIST_PLATFORM: 'linux-x64',
@@ -22,21 +23,22 @@ const envVars = {
 Object.entries(envVars).forEach(([key, val]) => envVars[key] = process.env[key] || val);
 
 const gitCommit = (() => {
-  const cmd = `git ls-remote ${STACKS_BLOCKCHAIN_REPO} ${envVars.STACKS_BLOCKCHAIN_BRANCH}`;
-  console.log(`Fetching commit hash for ${cmd}`);
-  const [commitHash] = execSync(cmd, {encoding: 'utf8'}).split(/(\s+)/);
+  console.log(`Fetching commit hash for ${envVars.STACKS_BLOCKCHAIN_REPO}#${envVars.STACKS_BLOCKCHAIN_BRANCH}`);
+  const [commitHash] = execFileSync('git', ['ls-remote', envVars.STACKS_BLOCKCHAIN_REPO, envVars.STACKS_BLOCKCHAIN_BRANCH], {encoding: 'utf8'}).split(/(\s+)/);
   return commitHash;
 })();
 const shortCommit = gitCommit.substr(0, 7);
 console.log(`Using commit hash: ${gitCommit}`);
 
-const [dateString] = new Date().toISOString().split('T');
+const dateString = new Date().toISOString()
+const [dateStringShort] = dateString.split('T');
 
-const releaseFileName = `${envVars.STACKS_BLOCKCHAIN_BIN}-${envVars.STACKS_BLOCKCHAIN_DIST_PLATFORM}-${dateString}-${shortCommit}`;
+const releaseFileName = sanitize([
+  envVars.STACKS_BLOCKCHAIN_BIN, envVars.STACKS_BLOCKCHAIN_DIST_PLATFORM, 
+  dateStringShort, shortCommit, envVars.STACKS_BLOCKCHAIN_BRANCH
+].join('-'), { replacement: '-' });
 console.log(`Upload dist file: '${releaseFileName}'`);
 
-const buildOutputDir = fs.mkdtempSync(path.join(__dirname, '.stacks-core-build-'));
-console.log(`Building stacks-blockchain binaries into '${buildOutputDir}'`);
 
 const storage = new Storage({
   projectId: 'ops-shared',
@@ -44,38 +46,41 @@ const storage = new Storage({
 });
 const bucket = storage.bucket(BUCKET_NAME);
 
-function buildDist() {
-  return new Promise((resolve, reject) => {
-    const cargoCmd = `cargo install --git https://github.com/blockstack/stacks-blockchain.git --rev "${gitCommit}" --bin=blockstack-core --debug --root /build-out`;
-    const dockerRunCmd = `docker run -v "${buildOutputDir}:/build-out" rust:stretch ${cargoCmd}`;
-    console.log(`Running build via docker: ${dockerRunCmd}`);
-    const result = spawn(dockerRunCmd, {
-      cwd: __dirname,
-      shell: '/bin/bash',
-    });
-    readline.createInterface(result.stderr).on('line', console.error);
-    readline.createInterface(result.stdout).on('line', console.log);
-    result.on('error', (error) => reject(error))
-    result.on('exit', (code, signal) => {
-      try {
-        if (code !== 0) {
-          const msg = `Cargo bad exit code: ${code}, signal: ${signal}`;
-          console.error(msg);
-          process.exit(code);
-        }
-      }
-      finally {
-        const binFileResult = `${buildOutputDir}/bin/blockstack-core`;
-        console.log(`Build complete: ${binFileResult}`);
-        resolve(binFileResult);
-      }
-    });
+const buildDist = () => new Promise((resolve, reject) => {
+  // const buildOutputDir = fs.mkdtempSync(path.join(__dirname, '.stacks-core-build-'));
+  const buildOutputDir = path.join(__dirname, `.stacks-core-build-${shortCommit}`);
+  const binFileResult = `${buildOutputDir}/bin/blockstack-core`;
+  fs.mkdirSync(buildOutputDir, { recursive: true });
+  if (fs.existsSync(binFileResult)) {
+    console.warn(`Skipping build step -- already exists at ${binFileResult}`);
+    resolve(binFileResult);
+    return;
+  }
+  console.log(`Building stacks-blockchain binaries into '${buildOutputDir}'`);
+  const cargoCmd = `cargo install --git https://github.com/blockstack/stacks-blockchain.git --rev "${gitCommit}" --bin=blockstack-core --debug --root /build-out`;
+  const dockerRunCmd = `docker run -v "${buildOutputDir}:/build-out" rust:stretch ${cargoCmd}`;
+  console.log(`Running build via docker: ${dockerRunCmd}`);
+  const result = spawn(dockerRunCmd, {
+    cwd: __dirname,
+    shell: '/bin/bash',
   });
-}
+  readline.createInterface(result.stderr).on('line', console.error);
+  readline.createInterface(result.stdout).on('line', console.log);
+  result.on('error', (error) => reject(error))
+  result.on('exit', (code, signal) => {
+    if (code !== 0) {
+      const msg = `Cargo bad exit code: ${code}, signal: ${signal}`;
+      console.error(msg);
+      process.exit(code);
+    }
+    console.log(`Build complete: ${binFileResult}`);
+    resolve(binFileResult);
+  });
+});
 
 const getBucketFileUrl = (file) => BUCKET_URL + file;
-const fetchBucketFile = (file) => new Promise((resolve, reject) => {
-  const request = https.get(getBucketFileUrl(file), (response) => {
+const fetchFile = (url) => new Promise((resolve, reject) => {
+  const request = https.get(url, (response) => {
     if (response.statusCode < 200 || response.statusCode > 299) {
         reject(new Error('Failed to load page, status code: ' + response.statusCode));
       }
@@ -95,16 +100,20 @@ async function uploadBucketFile(distFilePath) {
     gzip: true,
     contentType: 'application/octet-stream',
   });
+  const distFileUrl = getBucketFileUrl(releaseFileName);
+  console.log(`Uploaded ${distFileUrl}`);
 
   // Second, update the index.json pointers
   const releaseEntry = {
-    commit_hash: gitCommit,
+    git_branch: envVars.STACKS_BLOCKCHAIN_BRANCH,
+    git_commit_hash: gitCommit,
     date: dateString,
     file: releaseFileName,
     url: getBucketFileUrl(releaseFileName)
   };
   console.log('Fetching latest release index.json');
-  const currentIndex = JSON.parse(await fetchBucketFile('index.json'));
+  const indexFileUrl = getBucketFileUrl('index.json');
+  const currentIndex = JSON.parse(await fetchFile(indexFileUrl));
   const indexEntry = {
     [envVars.STACKS_BLOCKCHAIN_BIN]: {
       [envVars.STACKS_BLOCKCHAIN_DIST_PLATFORM]: {
@@ -124,7 +133,8 @@ async function uploadBucketFile(distFilePath) {
     metadata: { cacheControl: 'no-cache' }
   });
   await pipelineAsync(indexDataStream, writeStream);
-  console.log('release distribution completed');
+  console.log(`Updated ${indexFileUrl}`);
+  console.log('Release distribution completed');
 }
 
 async function run() {
